requirements.txt
langchain
langchain_community
python-dotenv
langchain-google-genai
langchain-openai
langchain-tavily